+++ camel/toolkits/__init__.py
@@ -12,7 +12,7 @@
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 # ruff: noqa: I001
-from .openai_function import (
+from .function_tool import (
     FunctionTool,
     OpenAIFunction,
     get_openai_function_schema,
